change the type of generator

+++ src/lambertian.h
@@ -6,7 +6,7 @@
 //generate a random floating point number to generate a slightly different direction for each ray
 inline double randomFloat()
 {
-    static std::uniform_real_distribution<float> dist(0.0,1.0);
+    static std::uniform_real_distribution<double> dist(0.0,1.0);
     static std::default_random_engine generator;
     return dist(generator);
 }
